+++ backend_v3/src/controllers/resourceController.ts
