+++ backend_v3/src/routes/resource.ts
@@ -1,12 +1,8 @@
 import express from 'express'
-import * as resourceController from '../controllers/resource.controller'
+import * as resourceController from '../controllers/resourceController'
 
 const router = express.Router()
 
-router.get(
-  '/:type',
-  resourceController.checkValidInput,
-  resourceController.getResource
-)
+router.get('/:type', resourceController.checkValidInput, resourceController.getResource)
 
 export default router
