+++ backend_v3/src/routes/vehicle.ts
@@ -1,5 +1,5 @@
 import express from 'express'
-import * as vehicleController from '../controllers/vehicle.controller'
+import * as vehicleController from '../controllers/vehicleController'
 
 const router = express.Router()
